requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
langchain==0.0.350
langchain-openai==0.0.2
langchain-groq==0.0.1
langchain-openrouter==0.0.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
python-docx==1.1.0
PyPDF2==3.0.1
reportlab==4.0.7
evaluate==0.4.1
tavily-python==0.3.3